refactoring specs and fix issue with unitspec not running
