+++ test/uk/gov/hmrc/brm/utils/FormatSpec.scala
@@ -14,10 +14,9 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.birthregistrationmatching.utility
+package uk.gov.hmrc.brm.utils
 
 import uk.gov.hmrc.play.test.UnitSpec
-
 import uk.gov.hmrc.utility._
 
 class FormatSpec extends UnitSpec {
