+++ test/uk/gov/hmrc/brm/utils/HeaderValidatorSpec.scala
@@ -17,11 +17,11 @@
 package uk.gov.hmrc.brm.utils
 
 import akka.stream.Materializer
-import com.google.inject.Inject
 import org.mockito.Matchers
 import org.mockito.Mockito._
 import org.scalatest.mock.MockitoSugar
 import org.scalatestplus.play.OneAppPerSuite
+import play.api.Play
 import play.api.libs.json.{JsValue, Json}
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
@@ -33,10 +33,7 @@ import uk.gov.hmrc.play.test.UnitSpec
 
 import scala.concurrent.Future
 
-class HeaderValidatorSpec @Inject()(implicit mat : Materializer) extends UnitSpec
-  with OneAppPerSuite
-  with MockitoSugar
-  with HeaderValidator {
+class HeaderValidatorSpec extends UnitSpec with OneAppPerSuite with MockitoSugar {
 
   private val jsonResponse = """{"code":"145","status":"400","details":"The headers you supplied are invalid","title":"Headers invalid","about":"http://http://htmlpreview.github.io/?https://github.com/hmrc/birth-registration-matching/blob/master/api-documents/api.html"}"""
 
@@ -49,9 +46,13 @@ class HeaderValidatorSpec @Inject()(implicit mat : Materializer) extends UnitSpe
     override val matchingService = MatchingService
   }
 
+  implicit lazy val materializer = Play.current.injector.instanceOf[Materializer]
+
   object MockController extends BirthEventsController {
     override val service = MockLookupService
   }
+
+  object HeaderValidator extends HeaderValidator
 
   val groJsonResponseObject = JsonUtils.getJsonFromFile("500035710")
 
@@ -71,35 +72,35 @@ class HeaderValidatorSpec @Inject()(implicit mat : Materializer) extends UnitSpe
   "acceptHeaderValidationRules" should {
 
     "return false when argument values are missing" in {
-      acceptHeaderValidationRules() shouldBe false
+      HeaderValidator.acceptHeaderValidationRules() shouldBe false
     }
 
     "return false when Accept header is invalid" in {
-      acceptHeaderValidationRules(accept = Some("text/html"), auditSource = Some("DFS")) shouldBe false
+      HeaderValidator.acceptHeaderValidationRules(accept = Some("text/html"), auditSource = Some("DFS")) shouldBe false
     }
 
     "return false when version is invalid" in {
-      acceptHeaderValidationRules(accept = Some("application/json"), auditSource = Some("DFS")) shouldBe false
+      HeaderValidator.acceptHeaderValidationRules(accept = Some("application/json"), auditSource = Some("DFS")) shouldBe false
     }
 
     "return false when auditSource is invalid" in {
-      acceptHeaderValidationRules(accept = Some("application/vnd.hmrc.1.0+json"), auditSource = Some("")) shouldBe false
+      HeaderValidator.acceptHeaderValidationRules(accept = Some("application/vnd.hmrc.1.0+json"), auditSource = Some("")) shouldBe false
     }
 
     "return true when Accept header and auditSource header are valid" in {
-      acceptHeaderValidationRules(accept = Some("application/vnd.hmrc.1.0+json"), auditSource = Some("DFS")) shouldBe true
+      HeaderValidator.acceptHeaderValidationRules(accept = Some("application/vnd.hmrc.1.0+json"), auditSource = Some("DFS")) shouldBe true
     }
 
     "return true when Accept header for mixed case and auditSource header are valid " in {
-      acceptHeaderValidationRules(accept = Some("application/vNd.HMRC.1.0+jSon"), auditSource = Some("DFS")) shouldBe true
+      HeaderValidator.acceptHeaderValidationRules(accept = Some("application/vNd.HMRC.1.0+jSon"), auditSource = Some("DFS")) shouldBe true
     }
 
     "return false when Accept header is not valid and auditSource header is valid " in {
-      acceptHeaderValidationRules(accept = Some(""), auditSource = Some("DFS")) shouldBe false
+      HeaderValidator.acceptHeaderValidationRules(accept = Some(""), auditSource = Some("DFS")) shouldBe false
     }
 
     "return false when Accept header has no value and auditSource header is valid " in {
-      acceptHeaderValidationRules(accept = None, auditSource = Some("DFS")) shouldBe false
+      HeaderValidator.acceptHeaderValidationRules(accept = None, auditSource = Some("DFS")) shouldBe false
     }
   }
 
@@ -110,7 +111,7 @@ class HeaderValidatorSpec @Inject()(implicit mat : Materializer) extends UnitSpe
           .withHeaders((ACCEPT, "application/vnd.hmrc.1.0+json"), ("Audit-Source", "DFS"))
           .withBody(payload)
         when(mockConnector.getReference(Matchers.any())(Matchers.any())).thenReturn(Future.successful(httpResponse(groJsonResponseObject)))
-        val result = MockController.post().apply(request)
+        val result = await(MockController.post().apply(request))
         status(result) shouldBe OK
       }
     }
@@ -122,7 +123,7 @@ class HeaderValidatorSpec @Inject()(implicit mat : Materializer) extends UnitSpe
           .withBody(payload)
         val result = await(MockController.post().apply(request))
         status(result) shouldBe BAD_REQUEST
-        jsonBodyOf(result) shouldBe jsonResponse
+        bodyOf(result) shouldBe jsonResponse
       }
     }
 
@@ -133,7 +134,7 @@ class HeaderValidatorSpec @Inject()(implicit mat : Materializer) extends UnitSpe
           .withBody(payload)
         val result = await(MockController.post().apply(request))
         status(result) shouldBe BAD_REQUEST
-        jsonBodyOf(result) shouldBe jsonResponse
+        bodyOf(result) shouldBe jsonResponse
       }
     }
 
@@ -144,7 +145,7 @@ class HeaderValidatorSpec @Inject()(implicit mat : Materializer) extends UnitSpe
           .withBody(payload)
         val result = await(MockController.post().apply(request))
         status(result) shouldBe BAD_REQUEST
-        jsonBodyOf(result) shouldBe jsonResponse
+        bodyOf(result) shouldBe jsonResponse
       }
     }
 
@@ -155,7 +156,7 @@ class HeaderValidatorSpec @Inject()(implicit mat : Materializer) extends UnitSpe
           .withBody(payload)
         val result = await(MockController.post().apply(request))
         status(result) shouldBe BAD_REQUEST
-        jsonBodyOf(result) shouldBe jsonResponse
+        bodyOf(result) shouldBe jsonResponse
       }
     }
 
@@ -166,7 +167,7 @@ class HeaderValidatorSpec @Inject()(implicit mat : Materializer) extends UnitSpe
           .withBody(payload)
         val result = await(MockController.post().apply(request))
         status(result) shouldBe BAD_REQUEST
-        jsonBodyOf(result) shouldBe jsonResponse
+        bodyOf(result) shouldBe jsonResponse
       }
     }
